backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.16.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.10.6
pydantic-settings==2.8.1
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.2.0
pytest==8.4.2
pytest-cov==7.0.0
